requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
cffi==1.12.3
chardet==3.0.4
discord.py==1.2.3
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
pkg-resources==0.0.0
pycparser==2.19
PyNaCl==1.3.0
six==1.12.0
typing-extensions==3.7.4
websockets==6.0
yarl==1.3.0
youtube-dl==2019.7.14
